Assume UTF-8 encoding for .markdown and .haml
New Tools.fixEncoding function can be used to adjust the result of
Git.readFile whenever we expect UTF-8 encoded file.

Should be pretty safe since ASCII is UFT-8
The real culprit here is git-fs which reads files to string using 'binary' encoding
We convert the file content back to buffer and re-encode it as UTF-8

It would be better if git-fs were fixed to read files content into buffer (and not into
string as it does now).